refactor(checker): route const-assert return widening through query_boundaries::widening
Move the object-shape read and shape-preserving rebuild wrappers from the
broad query_boundaries::common barrel into the concern-specific
query_boundaries::widening module, matching the const-assertion/widening
helpers already there. Keeps the query_boundaries::common reference count at
its #8225 cap (arch-ratchet) instead of adding two new direct barrel
references in return_type.rs.
