+++ crates/tsz-checker/src/query_boundaries/widening.rs
@@ -4,8 +4,8 @@
 //! `tsz_solver::*` directly (architecture rule: no inline solver function
 //! calls in checker modules).
 
-use tsz_solver::TypeId;
 use tsz_solver::construction::TypeDatabase;
+use tsz_solver::{ObjectShape, PropertyInfo, TypeId};
 
 /// Widen a type for inference resolution: deep-widens fresh literals while
 /// preserving function/callable parameter and return types unchanged.
@@ -43,4 +43,33 @@ pub(crate) fn apply_const_assertion(db: &dyn TypeDatabase, type_id: TypeId) -> T
 pub(crate) fn is_plain_object_or_array_shape(db: &dyn TypeDatabase, type_id: TypeId) -> bool {
     tsz_solver::type_queries::is_object_type(db, type_id)
         || tsz_solver::type_queries::is_array_or_tuple_type(db, type_id)
+}
+
+/// The [`ObjectShape`] of `type_id` when it is an object type, else `None`.
+///
+/// Exposed alongside the const-assertion widening helpers so the return-type
+/// inference path can read an object literal's shape to widen a subset of its
+/// properties (preserving const-asserted leaves) through this narrow boundary.
+pub(crate) fn object_shape_for_type(
+    db: &dyn TypeDatabase,
+    type_id: TypeId,
+) -> Option<std::sync::Arc<ObjectShape>> {
+    tsz_solver::type_queries::get_object_shape(db, type_id)
+}
+
+/// Rebuild an object type from `original` with `new_props`, preserving index
+/// signatures, flags (including `FRESH_LITERAL`), declaring symbol, and display
+/// provenance. Returns `original` unchanged when the properties are unchanged.
+///
+/// Pairs with [`object_shape_for_type`] for AST-driven partial widening that
+/// must keep the object's freshness/identity metadata.
+pub(crate) fn rebuild_object_with_shape_metadata(
+    db: &dyn TypeDatabase,
+    original: TypeId,
+    shape: &ObjectShape,
+    new_props: Vec<PropertyInfo>,
+) -> TypeId {
+    tsz_solver::operations::widening::rebuild_object_with_shape_metadata(
+        db, original, shape, new_props,
+    )
 }
